+++ src/chai.spec.ts
@@ -9,3 +9,7 @@ export const pfs = pify(require("fs"))
 export const ptmp = pify(require("tmp"))
 
 require("source-map-support").install()
+
+export function times<T>(n: number, f: ((idx: number) => T)): T[] {
+  return Array(n).fill(undefined).map((_, i) => f(i))
+}